% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LightLogWeb-app.R
\name{LightLogWeb}
\alias{LightLogWeb}
\title{Bring LightLogR to the web with shiny}
\usage{
LightLogWeb(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
Open a viewer with the shiny app
}
\description{
Bring LightLogR to the web with shiny
}
\examples{
if(interactive()) {
LightLogWeb()
}
}
