% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llw-data.R
\name{llw_merge}
\alias{llw_merge}
\title{Merge compatible LightLogWeb datasets}
\usage{
llw_merge(x, y, name = paste(x$name, y$name, sep = " + "))
}
\arguments{
\item{x, y}{\code{llw_dataset} objects.}

\item{name}{Name of the merged dataset.}
}
\value{
A merged \code{llw_dataset}.
}
\description{
Merge compatible LightLogWeb datasets
}
