% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llw-recipe.R
\name{llw_recipe}
\alias{llw_recipe}
\title{Create a LightLogWeb analysis recipe}
\usage{
llw_recipe(steps = list())
}
\arguments{
\item{steps}{A list of recipe steps. Most users build recipes with
\code{llw_recipe_add()}.}
}
\value{
An object of class \code{llw_recipe}.
}
\description{
Create a LightLogWeb analysis recipe
}
