% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llw-recipe.R
\name{llw_recipe_move}
\alias{llw_recipe_move}
\title{Move a recipe step}
\usage{
llw_recipe_move(recipe, id, direction = c("up", "down"))
}
\arguments{
\item{recipe}{An \code{llw_recipe}.}

\item{id}{Step ID.}

\item{direction}{Either \code{"up"} or \code{"down"}.}
}
\value{
An updated recipe.
}
\description{
Move a recipe step
}
