% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llw-recipe.R
\name{llw_recipe_undo}
\alias{llw_recipe_undo}
\alias{llw_recipe_redo}
\title{Undo or redo recipe changes}
\usage{
llw_recipe_undo(recipe)

llw_recipe_redo(recipe)
}
\arguments{
\item{recipe}{An \code{llw_recipe}.}
}
\value{
An updated recipe.
}
\description{
Undo or redo recipe changes
}
