% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llw-recipe.R
\name{llw_recipe_update}
\alias{llw_recipe_update}
\alias{llw_recipe_remove}
\title{Modify recipe steps}
\usage{
llw_recipe_update(recipe, id, parameters = NULL, enabled = NULL, label = NULL)

llw_recipe_remove(recipe, id)
}
\arguments{
\item{recipe}{An \code{llw_recipe}.}

\item{id}{Step ID.}

\item{parameters}{Replacement parameter list.}

\item{enabled}{Replacement enabled state.}

\item{label}{Replacement label.}
}
\value{
An updated recipe.
}
\description{
Modify recipe steps
}
