% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llw-modules-run.R
\name{llw_run}
\alias{llw_run}
\title{Execute a complete LightLogWeb recipe}
\usage{
llw_run(x, recipe = llw_recipe(), modules = list())
}
\arguments{
\item{x}{An \code{llw_dataset} or compatible data frame.}

\item{recipe}{An \code{llw_recipe}.}

\item{modules}{Startup-provided module specifications.}
}
\value{
An \code{llw_analysis}.
}
\description{
Execute a complete LightLogWeb recipe
}
