% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llw-modules-run.R
\name{llw_run_module}
\alias{llw_run_module}
\title{Run a custom module from the console}
\usage{
llw_run_module(x, module, params = list())
}
\arguments{
\item{x}{An \code{llw_dataset}.}

\item{module}{An \code{llw_module} specification.}

\item{params}{Named module parameters.}
}
\value{
An \code{llw_result}.
}
\description{
Run a custom module from the console
}
